refactor cells
