+++ NavioiOS/NavioiOS/UserInterface/MapBoard/LikePlace/PlaceCardCell.swift
@@ -1,0 +1,100 @@
+//
+//  PlaceCardCell.swift
+//  NavioiOS
+//
+//  Created by 김민우 on 9/15/25.
+//
+import UIKit
+import Navio
+import Combine
+import MapKit
+import ToolBox
+
+
+
+class PlaceCardCell: UICollectionViewCell {
+    // MARK: body
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupUI()
+    }
+    private func setupUI() {
+        // 계층 형성
+        contentView.addSubview(containerView)
+        containerView.addSubview(placeImageView)
+        containerView.addSubview(titleLabel)
+        containerView.addSubview(subtitleLabel)
+      
+        // 오토레이아웃 설정
+        NSLayoutConstraint.activate([
+            containerView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            containerView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            containerView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            containerView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
+
+            
+            placeImageView.topAnchor.constraint(equalTo: containerView.topAnchor),
+            placeImageView.leadingAnchor.constraint(equalTo: containerView.leadingAnchor),
+            placeImageView.trailingAnchor.constraint(equalTo: containerView.trailingAnchor),
+            placeImageView.heightAnchor.constraint(equalTo: containerView.heightAnchor, multiplier: 0.65),
+            
+            titleLabel.topAnchor.constraint(equalTo: placeImageView.bottomAnchor, constant: 8),
+            titleLabel.leadingAnchor.constraint(equalTo: containerView.leadingAnchor, constant: 16),
+            titleLabel.trailingAnchor.constraint(equalTo: containerView.trailingAnchor, constant: -16),
+            
+            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 4),
+            subtitleLabel.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
+            subtitleLabel.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
+        ])
+    }
+    
+    private let containerView: UIView = {
+        let view = UIView()
+        view.backgroundColor = .secondarySystemBackground
+        view.layer.cornerRadius = 34
+        view.layer.shadowColor = UIColor.black.cgColor
+        view.layer.shadowOpacity = 0.1
+        view.layer.shadowOffset = CGSize(width: 0, height: 2)
+        view.layer.shadowRadius = 8
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+    private let placeImageView: UIImageView = {
+        let iv = UIImageView()
+        iv.contentMode = .scaleToFill
+        iv.clipsToBounds = true
+        iv.layer.cornerRadius = 34
+        iv.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
+        iv.translatesAutoresizingMaskIntoConstraints = false
+        return iv
+    }()
+    private let titleLabel: UILabel = {
+        let label = UILabel()
+        label.font = UIFont.boldSystemFont(ofSize: 20)
+        label.textColor = .label
+        label.textAlignment = .right
+        label.numberOfLines = 1
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    private let subtitleLabel: UILabel = {
+        let label = UILabel()
+        label.font = UIFont.systemFont(ofSize: 15)
+        label.textColor = .systemGray
+        label.textAlignment = .right
+        label.numberOfLines = 2
+        label.lineBreakMode = .byWordWrapping
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    // 외부에서 데이터를 받아 셀의 UI를 업데이트하는 함수
+    func configure(with likePlace: LikePlace) {
+        placeImageView.image = likePlace.image
+        titleLabel.text = likePlace.name
+        subtitleLabel.text = likePlace.address
+    }
+}
