+++ NavioiOS/NavioiOS/UserInterface/MapBoard/SearchPlace/SearchPlaceModelVC.swift
@@ -15,77 +15,6 @@ struct SearchItemData {
   let subtitle: String
 }
 
-// MARK: - SearchListCell
-// 역할: UITableView 안에 들어갈 개별 검색 결과 셀의 UI와 레이아웃 정의.
-class SearchListCell: UITableViewCell {
-  
-  private let itemImageView: UIImageView = {
-    let imageView = UIImageView()
-    imageView.contentMode = .scaleAspectFit
-    imageView.clipsToBounds = true
-    imageView.layer.cornerRadius = 25
-    imageView.backgroundColor = .systemGray5
-    imageView.tintColor = .systemBlue
-      imageView.translatesAutoresizingMaskIntoConstraints = false
-    return imageView
-  }()
-  
-  private let titleLabel: UILabel = {
-    let label = UILabel()
-    label.font = UIFont.systemFont(ofSize: 18, weight: .medium)
-    label.textColor = .label
-    label.numberOfLines = 1
-      label.translatesAutoresizingMaskIntoConstraints = false
-    return label
-  }()
-  
-  private let subtitleLabel: UILabel = {
-    let label = UILabel()
-    label.font = UIFont.systemFont(ofSize: 14)
-    label.textColor = .systemGray
-    label.numberOfLines = 1
-      label.translatesAutoresizingMaskIntoConstraints = false
-    return label
-  }()
-  
-  override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-    super.init(style: style, reuseIdentifier: reuseIdentifier)
-    setupUI()
-  }
-  
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-  
-    private func setupUI() {
-        contentView.addSubview(itemImageView)
-        contentView.addSubview(titleLabel)
-        contentView.addSubview(subtitleLabel)
-        
-        selectionStyle = .none
-        
-        NSLayoutConstraint.activate([
-            itemImageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20),
-            itemImageView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-            itemImageView.widthAnchor.constraint(equalToConstant: 50),
-            itemImageView.heightAnchor.constraint(equalToConstant: 50),
-            
-            titleLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 15),
-            titleLabel.leadingAnchor.constraint(equalTo: itemImageView.trailingAnchor, constant: 15),
-            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -20),
-            
-            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 2),
-            subtitleLabel.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
-            subtitleLabel.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
-        ])
-  }
-  
-  func configure(with data: SearchItemData) {
-    itemImageView.image = UIImage(systemName: data.imageName)
-    titleLabel.text = data.title
-    subtitleLabel.text = data.subtitle
-  }
-}
 
 // MARK: - SearchPlaceModalViewController
 class SearchPlaceModelVC: UIViewController {
